pyproject.toml
[project]
name = "teaching-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Dmitry Karasev", email = "dakarasev@edu.hse.ru" }
]
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi>=0.128.0",
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",
  "pydantic-settings>=2.12.0",
  "python-multipart>=0.0.21",
  "uvicorn[standard]>=0.40.0",
]

[project.optional-dependencies]
queue = [
  "celery[redis]>=5.5.0,<6.0",
  "redis>=5.0.0",
]

llama = [
  "llama-index-core>=0.14.12",
  "llama-index-embeddings-huggingface>=0.6.1",
  "llama-index-llms-huggingface>=0.6.1",
  "qdrant-client>=1.16.2",
  "sentence-transformers>=3.0.0",
  "transformers>=4.40.0",
  "torch>=2.2.0",
]

dev = [
  "pre-commit>=4.5.1",
]

[project.scripts]
teaching-assistant = "teaching_assistant:main"

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"
